cleanup old exporter
