+++ aldryn_forms/compat.py
@@ -3,3 +3,8 @@ try:
     from collections import OrderedDict
 except ImportError:
     from django.utils.datastructures import SortedDict as OrderedDict
+
+try:
+    from formtools.wizard.views import SessionWizardView
+except ImportError:
+    from django.contrib.formtools.wizard.views import SessionWizardView
